backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pytest==7.4.2
httpx==0.25.0
python-dotenv==1.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
